requirements.txt
crewai==0.27.0
openai==1.99.9
crewai-tools==0.1.7
python-redmine
python-dotenv
requests
